Added a script to delete squashed-merged branhces
When squash-merging pull requests on Github/Gitlab etc, the local branch is no longer recognised as --merged, since the commits has been squashed, and therefore gotten a new sha1. This command creates a temporary dangling squashed commit with git commit-tree. Then uses git cherry to check if the squashed commit has already been applied to the current branch. If so, it deletes the branch.